requirements.txt
streamlit>=1.45.0
pandas>=2.3.0
pdfplumber>=0.11.0
gspread>=6.0.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
python-dotenv>=1.0.0
tqdm>=4.66.0
reportlab>=4.0.0
requests>=2.28.0
